feat: Shifting all teacher view content to teacher list
